==> doc/2.H_spatial_ops.Rmd <==
---
title: Spatial operations<br><h2>Threatened Australians (threatened.org.au)
author: Gareth Kindler<br>The University of Queensland, Australia
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
    html_document:
        code_folding: show
---

I run this code on a HPC, it could run locally but would just take a while.

## Libraries

```{r, class.source = 'fold-hide'}
library(tidyverse)
library(sf)
library(jsonlite)
library(magrittr)
library(units)
```

## Import: Electoral and Threatened Species

`postcodes`:
Postal Areas (POAs) were retreived from the Australian Bureau of Statistics (ABS) [Australian Statistical Geography Standard (ASGS) Edition 3](https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/non-abs-structures/postal-areas).

```{r}
elects <- st_read(
    "/QRISdata/Q4107/threatened_australians/output/clean_data/elects_clean.gpkg"
)
elects_union <- st_read(
    "/QRISdata/Q4107/threatened_australians/output/clean_data/elects_union_clean.gpkg"
)
species <- st_read(
    "/QRISdata/Q4107/threatened_australians/output/clean_data/species_clean.gpkg"
)
species_union <- st_read(
    "/QRISdata/Q4107/threatened_australians/output/clean_data/species_union_clean.gpkg"
    )
postcodes <- st_read(
    "/QRISdata/Q4107/threatened_australians/data/POA_2021_AUST_GDA94_SHP/POA_2021_AUST_GDA94.shp"
)
```

## Table: Postcodes-elects

The Threatened Australians web app is designed around which threatened species are found within which electorates. A simple way to connect users with their electorate is to ask for their postcode and have a table that links the two. For the postcodes that cross two or more electorates, we implemented a intermediary step which requests the user to select their electorate from a list of available options.
We spatially intersected the postcodes and CED data and calulcated the areas of each. This will allow us to filter out slight overlaps later-on.

```{r}
postcodes_elects_tbl <- postcodes %>%
    select(
        POA_CODE21, geometry
    ) %>%
    filter(
        !POA_CODE21 %in% c("9494", "9797", "ZZZZ")
    ) %>%
    rename(
        POA_code = POA_CODE21
    ) %>%
    mutate(
        POA_area_sqkm = units::set_units(st_area(.), km^2) %>%
            as.numeric()
    ) %>%
    st_intersection(elects) %>%
    st_make_valid() %>%
    mutate(
        POA_elect_int_area_sqkm = units::set_units(st_area(.), km^2) %>%
            as.numeric()
    ) %T>%
    st_write(
        dsn = "/QRISdata/Q4107/threatened_australians/output/analysed_data/final/postcodes_elects_tbl.gpkg",
        layer = "postcodes_elects_tbl", append = FALSE, delete_dsn = TRUE
    )
```

## Table: Species-elects

The `species_elects_tbl` forms the basis of the major function of the web app. This table shows us which threatened species intersect with which electorates.
`percent_range_within` is the proportion of each species range within within each electorate. This allows us to filter for endemism later. For example, the Southern Corroboree Frog (*Pseudophryne corroboree*) is found exclusively in Eden-Monaro. The attribute for this taxa will be == `1`.

```{r}
species_elects_tbl <- species %>%
    st_intersection(elects) %>%
    st_make_valid() %>%
    mutate(
        species_intersect_area_sqkm = units::set_units(st_area(.), km^2) %>%
            as.numeric()
    ) %>%
    mutate(
        percent_range_within = species_intersect_area_sqkm / species_range_area_sqkm
    ) %T>%
    st_write(
        dsn = "/QRISdata/Q4107/threatened_australians/output/analysed_data/final/species_elects_tbl.gpkg",
        layer = "species_elects_tbl", append = FALSE, delete_dsn = TRUE
    )
```

The unionised species data is used in for a Conversation article plot and not for the web app. It essentially allows us to calculate what proportion of each electorate (and consequently the incumbent's political affiliation [such as party or independent]) is threatened species range.

```{r}
species_union_elects <- species_union %>%
    st_intersection(elects) %>%
    st_make_valid() %>%
    mutate(
        elects_intersect_area_sqkm = units::set_units(st_area(.), km^2) %>%
            as.numeric()
    ) %T>%
    st_write(
        dsn = "/QRISdata/Q4107/threatened_australians/output/analysed_data/final/species_union_elects.gpkg",
        layer = "species_union_elects", append = FALSE, delete_dsn = TRUE
    )
```
