==> doc/3.l_electoral_clean.Rmd <==
---
title: Cleaning the electoral data<br><h2>Threatened Australians (threatened.org.au)
author: Gareth Kindler<br>The University of Queensland, Australia
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
    html_document:
        code_folding: show
---

```{r setup, include=FALSE, fig.align='center', warning=FALSE, message=FALSE}
# knitr::opts_chunk$set(echo = TRUE, comment = "#")
knitr::opts_knit$set(root.dir = "../")
```

## Libraries

```{r, message = FALSE, class.source = 'fold-hide'}
library(tidyverse)
library(sf)
library(magrittr)
library(jsonlite)
library(units)
library(readxl)
library(stringr)
library(rmapshaper)
library(httpgd)
```

## Import: Electoral

### Data key

`MP_info`:
Information about each Minister of Parliament (MP). Pulled from the Australian Parliament House API. Script for this can be found `~/scripts/MP_info/req_MP_info.py`

`MP_voting_info`:
Information about each MPs voting track record. Pulled from [They Vote For You](https://theyvoteforyou.org.au/) API. Script can be found at `~/scripts/MP_voting_info/req_MP_voting_info.py`

`demo`:
Demographic classification of Commonwealth Electoral Divisions (CEDs) was retrieved from the [Australian Electoral Commission (AEC)](https://www.aec.gov.au/Electorates/maps.htm).

```{r}
MP_info <- fromJSON(
    "data/22-12-13_MP_info.json"
)
MP_voting_info <- fromJSON(
    "data/22-12-13_MP_voting_info.json"
)
demo <- readxl::read_xlsx(
    "data/AEC_demographic-classification-2-august-2021/demographic-classification-as-at-2-august-2021.xlsx", skip = 3
)
elects <- st_read(
    "output/clean_data/elects_clean.gpkg"
)
```

## Clean: Electoral

### MP information

```{r}
MP_info_clean <- MP_info %>%
    mutate(
        electorate = word(
            .$representing, 1,
            sep = ","
        )
    ) |>
    mutate(
        image_URL = replace(
            image_URL, image_URL == "", "https://parlinfo.aph.gov.au/parlInfo/download/handbook/allmps/157125/upload_ref_binary/157125.JPG"
        )
    ) |>
    distinct() %>%
    relocate(
        electorate,
        .after = representing
    ) %T>%
    write_json(
        "output/clean_data/MP_info_clean.json"
    )
```

### MP voting information

We cleaned up the data of the Speaker of the House and the Chief Nationals Whip.

```{r}
MP_voting_info_clean <- MP_voting_info %>%
    filter(
        !is.na(electorate)
    ) |>
    mutate(
        party = replace(
            party, party == "SPK", "Australian Labor Party"
        )
    ) %>%
    mutate(
        party = replace(
            party, party == "CWM", "National Party"
        )
    ) %T>%
    write_json(
        "output/clean_data/MP_voting_info_clean.json"
    )
```

### Demographic classification of the CEDs

We cleaned up the data to be consistent and added in some acronyms.

```{r}
demo_clean <- demo %>%
    rename(
        state_territory = "State or territory",
        demographic_class = "Demographic classification",
        electorate = "Electoral division"
    ) %>%
    mutate(
        state_territory = replace(
            state_territory, state_territory == "ACT", "Australian Capital Territory"
        )
    ) %>%
    mutate(
        state_territory = replace(
            state_territory, state_territory == "NT", "Northern Territory"
        )
    ) %>%
    mutate(
        state_territory_abbrev = case_when(
            state_territory == "Australian Capital Territory" ~ "ACT",
            state_territory == "New South Wales" ~ "NSW",
            state_territory == "Northern Territory" ~ "NT",
            state_territory == "Queensland" ~ "QLD",
            state_territory == "South Australia" ~ "SA",
            state_territory == "Tasmania" ~ "TAS",
            state_territory == "Victoria" ~ "VIC",
            state_territory == "Western Australia" ~ "WA"
        )
    ) %T>%
    write_json(
        "output/clean_data/demo_clean.json"
    )
```
