==> doc/3.l_tsx_graphs.Rmd <==
---
title: Creating TSX data for graphs<br><h2>Threatened Australians (threatened.org.au)
author: Gareth Kindler<br>The University of Queensland, Australia
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
    html_document:
        code_folding: show
---

```{r setup, include=FALSE, fig.align='center', warning=FALSE, message=FALSE}
# knitr::opts_chunk$set(echo = TRUE, comment = "#")
knitr::opts_knit$set(root.dir = "../")
```

## Libraries

```{r}
library(tidyverse)
library(sf)
library(jsonlite)
library(magrittr)
library(httpgd)
```


##

Originally, we were recreating the index from the CSV the [interactive viewer](https://tsx.org.au/tsx/#/) was producing for each of the relevant states (NSW and ACT were kept together because of data paucity and geography) to get the trend lines. You can find the script for this in this file's history. Recently, the TSX peeps added the function to the interactive viewer that you can just download the trend lines for each state. Now we do this instead.


## Import: TSX

```{r}
tsx_NSW_ACT <- read.delim(
    "data/TSX/NSW_ACT_tsxdata/tsx-trend.txt",
    sep = " "
)
tsx_NT <- read.delim(
    "data/TSX/NT_tsxdata/tsx-trend.txt",
    sep = " "
)
tsx_QLD <- read.delim(
    "data/TSX/QLD_tsxdata/tsx-trend.txt",
    sep = " "
)
tsx_SA <- read.delim(
    "data/TSX/SA_tsxdata/tsx-trend.txt",
    sep = " "
)
tsx_TAS <- read.delim(
    "data/TSX/TAS_tsxdata/tsx-trend.txt",
    sep = " "
)
tsx_VIC <- read.delim(
    "data/TSX/VIC_tsxdata/tsx-trend.txt",
    sep = " "
)
tsx_WA <- read.delim(
    "data/TSX/WA_tsxdata/tsx-trend.txt",
    sep = " "
)
```

## Clean: TSX

```{r}
tsx_clean <- function(x) {
    x |>
        na.omit() |>
        rownames_to_column("ref_year") |>
        as_tibble()
}
```

```{r}
tsx_NSW_ACT_clean <- tsx_clean(tsx_NSW_ACT) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_NSW_ACT_tbl.json"
    )
tsx_NT_clean <- tsx_clean(tsx_NT) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_NT_tbl.json"
    )
tsx_QLD_clean <- tsx_clean(tsx_QLD) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_QLD_tbl.json"
    )
tsx_SA_clean <- tsx_clean(tsx_SA) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_SA_tbl.json"
    )
tsx_TAS_clean <- tsx_clean(tsx_TAS) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_TAS_tbl.json"
    )
tsx_VIC_clean <- tsx_clean(tsx_VIC) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_VIC_tbl.json"
    )
tsx_WA_clean <- tsx_clean(tsx_WA) %T>%
    write_json(.,
        "output/analysed_data/final/imports/tsx_WA_tbl.json"
    )
```
