==> doc/4.l_species_filters.Rmd <==
---
title: Filtering threatened species<br><h2>Threatened Australians (threatened.org.au)
author: Gareth Kindler<br>The University of Queensland, Australia
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
    html_document:
        code_folding: show
---

```{r setup, include=FALSE, fig.align='center', warning=FALSE, message=FALSE}
# knitr::opts_chunk$set(echo = TRUE, comment = "#")
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Libraries

```{r, message = FALSE, class.source = 'fold-hide'}
library(tidyverse)
library(sf)
library(magrittr)
library(jsonlite)
library(units)
library(readxl)
library(stringr)
library(rmapshaper)
library(httpgd)
```

## Import: Threatened Species and threats

```{r, message = FALSE}
species <- st_read(
    "output/clean_data/species_clean.gpkg"
)
species_elects <- st_read(
    "output/analysed_data/final/species_elects_tbl.gpkg"
)
threats_collapsed <- fromJSON(
    "output/clean_data/threats_collapsed_clean.json"
)
```

## Filter: Species

This is a dataframe used to filter (`ft`) others based on the final species we'll be including on the web app - those that intersect with electorates and have threat information.

```{r}
species_ft <- species_elects %>%
    st_set_geometry(NULL) %>%
    group_by(taxon_ID) %>%
    summarise() %>%
    ungroup() %>%
    # inner_join(threats_collapsed) %>%
    select(taxon_ID) %T>%
    write_json(
        "output/clean_data/species_ft.json"
    )
```

## Filter: Species

### Freshwater and terrestrial

The SNES data attributes of `marine` and `cetacean` don't capture all species of these categories so some manual cleaning needed to occur.

```{r}
species_animals_clean <- species %>%
    st_set_geometry(NULL) %>%
    # inner_join(species_ft) %>%
    filter(
        taxon_kingdom %in% "Animalia"
    ) %>%
    filter(
        !marine %in% c(
            "Listed", "Listed - overfly marine area"
        )
    ) %>%
    filter(
        !cetacean %in% "Cetacean"
    ) %>%
    filter(
        !scientific_name %in% c(
            "Brachionichthys hirsutus", # Spotted Handfish
            "Brachiopsilus ziebelli", # Ziebell's Handfish, Waterfall Bay Handfish
            "Carcharias taurus (east coast population)", # Grey Nurse Shark (east coast population)
            "Carcharias taurus (west coast population)", # Grey Nurse Shark (west coast population)
            "Carcharodon carcharias", # White Shark, Great White Shark
            "Epinephelus daemelii", # Black Rockcod, Black Cod, Saddled Rockcod
            "Glyphis garricki", # Northern River Shark, New Guinea River Shark
            "Glyphis glyphis", # Speartooth Shark
            "Pristis clavata", # Dwarf Sawfish, Queensland Sawfish
            "Rhincodon typus", # Whale Shark
            "Thymichthys politus", # Red Handfish
            "Zearaja maugeana", # Maugean Skate, Port Davey Skate
            "Thunnus maccoyii", # Southern Bluefin Tuna
            "Diomedea antipodensis gibsoni", # Gibson's Albatross, whack geom
            "Diomedea antipodensis", # Antipodean Albatross
            "Pachyptila turtur subantarctica", # Fairy Prion (southern)
            "Thalassarche salvini", # Salvin's Ablatross
            "Thalassarche steadi", # White-capped Albatross
            "Thalassarche eremita", # Chatham Albatross
            "Diomedea sanfordi", # Northern Royal Albatross
            "Diomedea epomophora", # Southern Royal Albatross
            "Sterna vittata bethunei",	# New Zealand Antarctic Tern
            "Sterna vittata vittata", # Antarctic Tern (Indian Ocean)
            "Leucocarbo atriceps nivalis",	# Imperial Shag (Heard Island)
            "Leucocarbo atriceps purpurascens", # Imperial Shag (Macquarie Island)
            "Centrophorus harrissoni",	# Harrisson's Dogfish
            "Galeorhinus galeus", # School Shark
            "Hoplostethus atlanticus",	# Orange Roughy
            "Seriolella brama", # Blue Warehou
            "Centrophorus zeehaani", #	Southern Dogfish
            "Dendronephthya australis" # Cauliflower Soft Coral
        )
    ) %T>%
    write_json(
        "output/clean_data/species_animals_clean.json"
    )
```

### Plants

```{r}
species_plants_clean <- species %>%
    st_set_geometry(NULL) %>%
    # inner_join(species_ft) %>%
    filter(
        taxon_kingdom %in% "Plantae"
    ) %T>%
    # filter(
    #     !marine %in% c(
    #         "Listed", "Listed - overfly marine area"
    #     )
    # ) %T>%
    write_json(
        "output/clean_data/species_plants_clean.json"
    )
```

### Marine

NB: Haven't updated this list to reflect the fish etc captured in animals as I've started to do. I'm not super interested in perfecting the segregation as I only care about getting animals relevant to the electoral system.

```{r}
species_marine_clean <- species %>%
    st_set_geometry(NULL) %>%
    # inner_join(species_ft) %>%
    filter(
        marine %in% c(
            "Listed", "Listed - overfly marine area"
        ) |
            cetacean %in% "Cetacean" |
            scientific_name %in% c(
                "Brachionichthys hirsutus", # Spotted Handfish
                "Brachiopsilus ziebelli", # Ziebell's Handfish, Waterfall Bay Handfish
                "Carcharias taurus (east coast population)", # Grey Nurse Shark (east coast population)
                "Carcharias taurus (west coast population)", # Grey Nurse Shark (west coast population)
                "Carcharodon carcharias", # White Shark, Great White Shark
                "Epinephelus daemelii", # Black Rockcod, Black Cod, Saddled Rockcod
                "Glyphis garricki", # Northern River Shark, New Guinea River Shark
                "Glyphis glyphis", # Speartooth Shark
                "Pristis clavata", # Dwarf Sawfish, Queensland Sawfish
                "Rhincodon typus", # Whale Shark
                "Thymichthys politus", # Red Handfish
                "Zearaja maugeana", # Maugean Skate, Port Davey Skate
                "Thunnus maccoyii" # Southern Bluefin Tuna
            )
    ) %T>%
    write_json(
        "output/clean_data/species_marine_clean.json"
    )
```
