==> doc/6.l_compiling.Rmd <==
---
title: Compiling the clean and analysed data into tables<br><h2>Threatened Australians (threatened.org.au)
author: Gareth Kindler<br>The University of Queensland, Australia
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
    html_document:
        code_folding: show
---

```{r setup, include=FALSE, fig.align='center', warning=FALSE, message=FALSE}
# knitr::opts_chunk$set(echo = TRUE, comment = "#")
knitr::opts_knit$set(root.dir = "../")
```

## Libraries

```{r, message = FALSE, class.source = 'fold-hide'}
library(tidyverse)
library(sf)
library(rmapshaper) # For installing use 'library(remotes)'
library(jsonlite)
library(magrittr)
library(units)
library(httpgd)
```

## Import: Electoral

Explanations of what each of these data represent can be found in `2.local_attributes_clean.r`

```{r}
elects <- st_read(
    "output/clean_data/elects_clean.gpkg"
)
postcodes_elects <- st_read(
    "output/analysed_data/final/postcodes_elects_tbl.gpkg"
)
MP_info <- fromJSON(
    "output/clean_data/MP_info_clean.json"
)
MP_voting_info <- fromJSON(
    "output/clean_data/MP_voting_info_clean.json"
)
demo <- fromJSON(
    "output/clean_data/demo_clean.json"
)
```

## Import: Threatened Species

Explanations of what each of these data represent can be found in `2.local_attributes_clean.r`

```{r}
species_elects <- st_read(
    "output/analysed_data/final/species_elects_tbl.gpkg"
)
species <- st_read(
    "output/clean_data/species_clean.gpkg"
)
threats <- fromJSON(
    "output/clean_data/threats_clean.json"
)
threats_collapsed <- fromJSON(
    "output/clean_data/threats_collapsed_clean.json"
)
animals_ft <- fromJSON(
    "output/clean_data/species_animals_clean.json"
) %>%
    select(taxon_ID)
plants_ft <- fromJSON(
    "output/clean_data/species_plants_clean.json"
) %>%
    select(taxon_ID)
animals_images <- fromJSON(
    "output/clean_data/animals_images_clean.json"
)
animals_info <- fromJSON(
    "output/clean_data/animals_info_clean.json"
)
```

## Import: Other

`action_groups` is a manually created table that details the environmental NGOs we want to highlight on the website.

```{r}
action_groups <- fromJSON(
    "data/action_groups.json"
)
```

## Table: Species-elects

The bulk of the following code chunk is dedicated to creating an `species_electorate_coverage` attribute. This is a string that we can showcase on each of the unique combinations of species-elects that occur (~3.2K). The string describes how unique the species is to the electorate(s).

```{r}
species_elects_tbl <- species_elects %>%
    mutate(
        across(
            c(
                species_range_area_sqkm,
                species_intersect_area_sqkm,
                percent_range_within
            ),
            signif,
            digits = 3
        )
    ) %>%
    group_by(
        taxon_ID
    ) %>%
    mutate(species_range_intersects_with_n_electorates = n_distinct(electorate)) %>%
    ungroup() %>%
    mutate(
        scientific_name_clean = word(
            scientific_name, 1, 1,
            sep = fixed("(")
        )
    ) %>%
    # filter(
    #     percent_range_within >= 0.05 # TODO: gotta solve this among others
    # )
    # This introduces various errors, but for most species it's going to be good
    mutate(
        species_electorate_coverage = case_when(
            percent_range_within == 1 ~ paste0(
                scientific_name_clean,
                " is only found within ",
                electorate
            ),
            percent_range_within >= 0.8 &
                percent_range_within < 1
            # species_range_intersects_with_n_electorates != 1
            ~ paste0(
                    scientific_name_clean,
                    " has greater than 80% of its range within ",
                    electorate
                ),
            TRUE ~ paste0(
                scientific_name_clean,
                " is found across ",
                species_range_intersects_with_n_electorates,
                " electorates."
            )
        )
    ) %>%
    mutate(
        taxon_ID = as.integer(
            taxon_ID
        )
    ) %>%
    mutate(
        vernacular_name_first_clean = replace(
            vernacular_name_first_clean,
            vernacular_name_first_clean == "cycad",
            "a cycad "
        )
    ) %>%
    mutate(
        vernacular_name_first_clean = replace(
            vernacular_name_first_clean,
            vernacular_name_first_clean == "fern",
            "a fern"
        )
    ) %>%
    mutate(
        vernacular_name_first_clean = replace(
            vernacular_name_first_clean,
            vernacular_name_first_clean == "bluegrass",
            "Bluegrass"
        )
    )
```

### Animals-elects

```{r}
animals_elects_tbl <- species_elects_tbl %>%
    inner_join(animals_ft) %>%
    select(
        taxon_ID, scientific_name_clean,
        species_range_intersects_with_n_electorates,
        species_range_area_sqkm, species_electorate_coverage, electorate,
        species_intersect_area_sqkm, percent_range_within, geom
    ) %T>%
    st_write(
        "output/analysed_data/final/imports/animals_elects_tbl.geojson",
        layer = "animals_elects_tbl", append = FALSE, delete_dsn = TRUE
    )
```

### Plants-elects

Two `vernacular_name`s of plants were missing indefinite articles which were creating some grammatical errors.

```{r}
plants_elects_tbl <- species_elects_tbl %>%
    st_set_geometry(NULL) %>%
    inner_join(plants_ft) %>%
    select(
        taxon_ID,
        scientific_name_clean,
        vernacular_name_first_clean,
        threatened_status,
        SPRAT_profile,
        electorate
    ) %T>%
    write_json(
        "output/analysed_data/final/imports/plants_elects_tbl.json"
    )
```

## Reference tables

### All-encompassing

This creates easily filter-able tables (no geometry) for use as reference. For example, someone wants to know only the critically endangered birds that occur across the Northern Territory or Lingiari - this tables enables them to do that.

```{r}
species_ft <- animals_ft %>%
    bind_rows(plants_ft)

species_elects_ref_tbl <- species_elects_tbl %>%
    st_set_geometry(NULL) %>%
    inner_join(species_ft) %>%
    mutate(
        across(
            percent_range_within,
            signif,
            digits = 3
        )
    ) %>%
    full_join(demo) %>%
    select(
        electorate, state_territory, electorate_area_sqkm,
        taxon_ID, scientific_name,
        # scientific_name_clean,
        vernacular_name,
        # vernacular_name_other, vernacular_name_first,
        # vernacular_name_first_clean,
        threatened_status,
        migratory_status, marine, taxon_group, taxon_kingdom,
        SPRAT_profile, species_range_area_sqkm,
        species_intersect_area_sqkm, percent_range_within
    ) %T>%
    write_csv(
        "output/analysed_data/ref_tables/species_elects_ref_tbl.csv"
    )
```

## Table: Animals

```{r}
animals_tbl <- species %>%
    right_join(animals_ft) |>
    full_join(animals_info) |>
    mutate(
        description = replace_na(
            description,
            "No information was found for this species on the Species Profile and Threats Database (SPRAT) website, which is the database designed to provide information on species listed as threatened under the Environment Protection and Biodiversity Conservation (EPBC) Act 1999. This does not mean there is no information out there. We encourage you to do a web search using the scientific latin name."
        )
    ) |>
    full_join(animals_images) |>
    left_join(threats_collapsed) |>
    mutate(
        threat_ID_collapsed = replace_na(
            threat_ID_collapsed,
            "T09"
        )
    ) |>
    right_join(animals_ft) |>
    mutate(
        vernacular_name_first_clean = replace(
            vernacular_name_first_clean,
            vernacular_name_first_clean == "freshwater crayfish",
            "a freshwater crayfish"
        )
    ) %>%
    mutate(
        vernacular_name_first_clean = replace(
            vernacular_name_first_clean,
            vernacular_name_first_clean == "A native bee",
            "a native bee"
        )
    ) %>%
    relocate(
        vernacular_name_first,
        .after = vernacular_name
    ) %>%
    relocate(
        vernacular_name_other,
        .after = vernacular_name_first
    ) %>%
    relocate(
        vernacular_name_first_clean,
        .after = vernacular_name_first
    ) %>%
    mutate(
        taxon_ID = as.integer(
            taxon_ID
        )
    ) %>%
    select(
        taxon_ID, scientific_name_clean, vernacular_name_first_clean,
        threatened_status, migratory_status, taxon_group,
        taxon_kingdom, SPRAT_profile,
        species_range_area_sqkm, description, ALA_URL, image_URL,
        threat_ID_collapsed, geom
    ) %>%
    mutate(
        across(
            species_range_area_sqkm,
            signif,
            digits = 3
        )
    ) %T>%
    st_write(
        "output/analysed_data/final/imports/animals_tbl.geojson",
        layer = "animals_tbl", append = FALSE, delete_dsn = TRUE
    )
```


```{r check-animals_tbl}
summary(animals_tbl)
apply(animals_tbl, MARGIN = 2, function(x) sum(is.na(x)))
get_range <- function(x) {
    if (is.numeric(x)) {
        return(paste0("min: ", min(x), ", max: ", max(x)))
    } else {
        return(paste0("unique: ", length(unique(x))))
    }
}
apply(animals_tbl, MARGIN = 2, get_range)
```

Based on the `scientific_name_clean` coming out as one less unique than the `taxon_ID` - `r animals_tbl |> group_by(scientific_name_clean) |> summarise(n = n()) |> filter(n > 1)`. The Spot-tailed quoll and the Spotted-tail quoll.

## Table: Elects

On the *Browser* page, there is the option to explore other electorates by a couple of filters, one such is by the number of animals and plants found within the boundaries of each electorate.

```{r}
animals_elects_counts <- species_elects %>%
    st_set_geometry(NULL) %>%
    inner_join(animals_ft) %>%
    group_by(electorate) %>%
    summarise(no_animal_species = n_distinct(taxon_ID)) %>%
    ungroup()

plants_elects_counts <- species_elects %>%
    st_set_geometry(NULL) %>%
    inner_join(plants_ft) %>%
    group_by(electorate) %>%
    summarise(no_plant_species = n_distinct(taxon_ID)) %>%
    ungroup()

elects_tbl <- elects %>%
    inner_join(demo) %>%
    inner_join(MP_info) %>%
    inner_join(MP_voting_info) %>%
    inner_join(animals_elects_counts) %>%
    full_join(plants_elects_counts) %>% # Not all elects have plants
    mutate(
        no_plant_species = replace_na(
            no_plant_species, 0
        )
    ) %>%
    mutate(
        across(
            electorate_area_sqkm,
            signif,
            digits = 3
        )
    ) %T>%
    st_write(
        "output/analysed_data/final/imports/elects_tbl.geojson",
        layer = "elects_tbl", append = FALSE, delete_dsn = TRUE
    ) |>
    st_set_geometry(NULL) %T>%
    write_csv(
        "output/analysed_data/ref_tables/elects_tbl.csv"
    )
```

```{r check-elects_tbl}
summary(elects_tbl)
apply(elects_tbl, MARGIN = 2, function(x) sum(is.na(x)))
get_range <- function(x) {
    if (is.numeric(x)) {
        return(paste0("min: ", min(x), ", max: ", max(x)))
    } else {
        return(paste0("unique: ", length(unique(x))))
    }
}
apply(elects_tbl, MARGIN = 2, get_range)
```

## Table: Postcodes-elects

As alluded to in `3.HPC_spatial_ops.Rmd`, intersecting `postcodes` and `elects` will cause intersections to be produced where there is slight overlaps, to an excessive degree. The slight overlaps are miniscule and often too small to be residential areas. We dealt with this by filtering out at the proportion of POA found within each electorate when it is less or equal to `0.001`.

```{r}
postcodes_elects_tbl <- postcodes_elects %>%
    st_set_geometry(NULL) %>%
    mutate(
        across(
            c(
                POA_area_sqkm, POA_elect_int_area_sqkm
            ),
            signif,
            digits = 3
        )
    ) %>%
    mutate(
        proportion_POA_within_elect = POA_elect_int_area_sqkm / POA_area_sqkm
    ) %>%
    filter(
        !proportion_POA_within_elect <= 0.001
    ) %T>%
    write_json(
        "output/analysed_data/final/imports/postcodes_elects_tbl.json"
    ) %T>%
    write_csv(
        "output/analysed_data/ref_tables/postcodes_elects_ref_tbl.csv"
    )
```

## Table: Action-groups

`Action_groups` just needs to be passed into the data import folder.

```{r}
action_groups_tbl <- action_groups %T>%
    write_json(
        "output/analysed_data/final/imports/action_groups_tbl.json"
    )
```

## Table: Threats

As the `species_tbl` has the `threat_ID` for the broad level threats each species faces, we just need a linking table for more information.

```{r}
threats_tbl <- threats %>%
    group_by(
        broad_level_threat, threat_ID
    ) %>%
    summarise() %>%
    relocate(
        threat_ID,
        .before = broad_level_threat
    ) %T>%
    write_json(
        "output/analysed_data/final/imports/threats_tbl.json"
    )
```
